requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arxiv==2.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.31
openai==1.109.1
orjson==3.11.3
packaging==25.0
pydantic==2.11.9
pydantic-core==2.33.2
python-dotenv==1.1.1
pyyaml==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.43
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
zstandard==0.25.0
